\name{SampleFunc}
\alias{SampleFunc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Simulate longitudinal data
}
\description{
	Generates a longitudinal dataset for 2-5 clusters.
}
\usage{
	SampleFunc(N.sub.clust=50, N.clust=2, N.tp=10, max.tp=10, min.tp=2, int.tp=c(0, 1), sd=0.2, reg=TRUE, dformat=1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	\item{N.sub.clust}{
		Sample size for each cluster.
	}
	\item{N.clust}{
		Number of clusters (between 1 and 5).
	}
	\item{N.tp}{
		Number of time points for regular datasets.
	}
	\item{max.tp}{
		Maximum number of time points for irregular datasets.
	}
	\item{min.tp}{
		Minimum number of time points for irregular datasets.
	}
	\item{int.tp}{
		Time interval from which the time points are drawn. Must be specified as a vector of length 2.
	}
	\item{sd}{
		Standard deviation from the cluster centers.
	}
	\item{reg}{
		If \code{TRUE}, generates regular datasets; if \code{FALSE}, generates irregular datasets.
	}
	\item{dformat}{
		Data format; 1 for list format, 2 for data frame format. For more details, please refer to the description of the \code{fdata} argument in the help documentation for the \code{fkms} function.
	}
}
\details{
	Curves are generated by adding a normally distributed error term with mean 0 and standard deviation \code{sd} to the center functions. The center functions are sampled from
	\deqn{
		x^2\\
		\sqrt{x}\\
		\sin(2\pi x)\\
		x^3
		-x^2
	}
	if \code{reg}=\code{TRUE}, all curves are evaluated on the same time points. If \code{reg}=\code{FALSE}, evaluation place and number can differ for each subject.
}
\value{
	\item{data}{
		The generated dataset.
	}	
	\item{cls}{
		The cluster number to which each subject belongs.
	}	
}
% \references{
% 	Yamamoto, M. and Terada, Y. (2024). K-means clustering for sparsely observed longitudinal data. submitted.
% }
\author{
	Michio Yamamoto\cr
	\email{yamamoto.michio.hus@osaka-u.ac.jp}
}
% \note{
% %%  ~~further notes~~
% }

% %% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
	\code{\link{fkms}}
}
\examples{
	## Example data with dformat=1 (list format)
	dat <- SampleFunc(N.clust=2, reg=FALSE, dformat=1)
	fdata <- dat$data ## sample curves
	cls.t <- dat$cls ## true clusters

	## Example data with dformat=2 (data.frame format)
	dat <- SampleFunc(N.clust=3, reg=FALSE, dformat=2)
	fdata <- dat$data ## sample curves
	cls.t <- dat$cls ## true clusters
}

